capstone-project-9900h14akuangbiao-main/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cssselect==1.2.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-filter==22.1
django-jazzmin==2.6.0
django-rest-swagger==2.2.0
django-taggit==3.1.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
drf-yasg2==1.19.4
filelock==3.10.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
parsel==1.7.0
Pillow==9.4.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.6.0
pyOpenSSL==23.0.0
python3-openid==3.2.0
pytz==2022.7.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Scrapy==2.8.0
scrapy-djangoitem==1.1.1
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
# twisted-iocpsupport
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
utils==1.0.1
w3lib==2.1.1
zope.interface==6.0
